requirements.txt
blinker==1.8.2
click==8.1.7
flask==2.3.2
Flask-Cors==4.0.1
flask-sqlalchemy==3.0.3
greenlet==3.0.3
gunicorn==22.0.0
importlib-metadata==7.1.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
SQLAlchemy==2.0.30
typing-extensions==4.12.2
werkzeug==3.0.3
zipp==3.19.2
psycopg2-binary==2.9.6
python-dotenv==0.21.0
Flask-Migrate
Flask-Bcrypt
Flask-Login
Flask-Script
